requirements.txt
numpy==1.18.2
opencv-python==4.2.0.32
Pillow==7.0.0
